+++ R/seurat.R
@@ -355,6 +355,11 @@ Images <- function(object, assay = NULL) {
 #'
 LoadSeuratRds <- function(file, ...) {
   object <- readRDS(file = file, ...)
+  rdsdir <- if (is_bare_character(x = file, n = 1L)) {
+    dirname(path = normalizePath(path = file, winslash = '/', mustWork = FALSE))
+  } else {
+    NULL
+  }
   cache <- Tool(object = object, slot = 'SaveSeuratRds')
   reqd.cols <- c('layer', 'path', 'class', 'pkg', 'fxn', 'assay')
   emit <- ifelse(
@@ -384,6 +389,16 @@ LoadSeuratRds <- function(file, ...) {
       emit(message = "Incorrect layer cache: none of the assays listed present")
       return(object)
     }
+    # Resolve the cached paths relative to the rds file; this keeps objects
+    # loadable after the rds and its on-disk stores have been moved together,
+    # and when relative paths are loaded from a different working directory
+    cache$path <- vapply(
+      X = cache$path,
+      FUN = .ResolveLayerPaths,
+      FUN.VALUE = character(length = 1L),
+      USE.NAMES = FALSE,
+      dir = rdsdir
+    )
     # Check the files
     exists <- vapply(
       X = cache$path,
@@ -770,10 +785,9 @@ SaveSeuratRds <- function(
           class = 'sticky',
           amount = 0
         )
-        df[i, 'path'] <- as.character(x = .FileMove(
-          path = pth,
-          new_path = destdir
-        ))
+        # a layer joined from several on-disk matrices records all of their
+        # paths in one entry, and each of them has to be moved
+        df[i, 'path'] <- .FileMoveAll(path = pth, new_path = destdir)
       }
     }
     if (isTRUE(x = relative)) {